make printf format and argument signedness match (#1082)

+++ common/src/arch-x86.C
@@ -9429,7 +9429,7 @@ static unsigned int ia32_decode_modrm(const unsigned int addrSzAttr, const unsig
 void ia32_memacc::print()
 {
     fprintf(stderr, "base: %d, index: %d, scale:%d, disp: %ld (%lx), size: %d, addr_size: %d\n",
-	    regs[0], regs[1], scale, imm, imm, size, addr_size);
+	    regs[0], regs[1], scale, imm, (unsigned long)imm, size, addr_size);
 }
 
 int getOperSz(const ia32_prefixes &pref)
